Nieuwe service klasse aangemaakt en AthleteEntry en NOCRegion naar aparte klassen gebracht
