+++ src/app/csv.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { CsvService } from './csv.service';
+
+describe('CsvService', () => {
+  let service: CsvService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(CsvService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
