+++ src/app/csv.service.ts
@@ -1,0 +1,10 @@
+import { HttpClient } from '@angular/common/http';
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class CsvService {
+
+  constructor(private http: HttpClient) { }
+}
